+++ packages/web3-eth/src/utils/prepare_transaction_for_signing.ts
@@ -134,12 +134,16 @@ export const prepareTransactionForSigning = async (
 		fillGasPrice,
 		fillGasLimit,
 	})) as unknown as PopulatedUnsignedTransaction;
-	const formattedTransaction = formatTransaction(
-		populatedTransaction,
-		ETH_DATA_FORMAT,
-	) as unknown as FormatType<PopulatedUnsignedTransaction, typeof ETH_DATA_FORMAT>;
+	const formattedTransaction = formatTransaction(populatedTransaction, ETH_DATA_FORMAT, {
+		transactionSchema: web3Context.config.customTransactionSchema,
+	}) as unknown as FormatType<PopulatedUnsignedTransaction, typeof ETH_DATA_FORMAT>;
+
 	validateTransactionForSigning(
 		formattedTransaction as unknown as FormatType<Transaction, typeof ETH_DATA_FORMAT>,
+		undefined,
+		{
+			transactionSchema: web3Context.config.customTransactionSchema,
+		},
 	);
 
 	return TransactionFactory.fromTxData(